update api  documentation
