+++ README.md
@@ -16,7 +16,7 @@ import ravepaypysdk
 my_api = ravepaypysdk.Api(
           SECRET_KEY='ravepay_secret_key'
           PUBLIC_KEY='ravepay_public_key'
-          Production=False # sandbox # or True # Live
+          production=False # sandbox # or True # Live
         )
 ```
 
@@ -83,94 +83,172 @@ To start using the SDK ensure you have your public key and secret key instantiat
         return bank_payment
 
         ```
-  - The same format goes for the **USSD**, **MPESA** and **Ghana Mobile Money**. Kindly review the API documentation to get the required field for each transaction
+  -  **Tokenize Card**
+
+     *Usage*
+     ```
+     payload = {
+       "token":{
+           “chargeToken”:{
+                “user_token”:“f0209”,
+                “embed_token”:“flw-t0-9f3aa69a806f6440fbb78cc9e8b2f135-k3n”
+            }
+       },
+       "currency": "NGN",
+       "country": "NG",
+       "amount": "10",
+       "email": "user@example.com",
+       "phonenumber": "1234555",
+       "firstname": "user1",
+       "lastname": "user2",
+       "IP": "355426087298442",
+       "txRef": "MC-7663-YU",
+       "narration":"tokenize my card"
+       "meta": [{"metaname": "flightId"}, {"metavalue": "1002"}]
+       "device_fingerprint": "69e6b7f0b72037aa8428b70fbe03986c"
+     }
+
+     tokenize_card = Payment.tokenize_card(payload, api=my_api)
+     ```
+  -  **USSD Payment**
+
+     *Usage*
+     ```
+     ussd_payment = Payment.ussd(payload, api=my_api)
+     ```
+
+  -  **MPESA Payment**
+
+     *Usage*
+     ```
+     mpesa_payment = Payment.mpesa(payload, api=my_api)
+     ```
+
+  -  **Ghana Mobile Money**
+
+     *Usage*
+     ```
+     gh_mobile_payment = Payment.ghana_mobile(payloadm api=my_api)
+     ```
+
+  - The same payload format goes for the **USSD**, **MPESA** and **Ghana Mobile Money**. Kindly review the [API documentation](https://flutterwavedevelopers.readme.io/v1.0/reference#rave-parameters) to get the required field for each transaction
 
 ## Transaction
-    Import the transaction module
-    ```
-    from ravepaypysdk import Transaction
-    ```
+  This module retrieves all transactions and verifies transactions.
 
-  - **Verify transaction**
+  *Usage*
+
+  Import the transaction module
+  
+      from ravepaypysdk import Transaction
+    
+
+  - **Verify your transaction**
 
       *Usage*
       ```
+      payload = {
+        "flw_ref": "FLW-MOCK-6f52518a2ecca2b6b090f9593eb390ce", # unique reference for the transaction
+        "tx_ref":"dummy", # merchants unique reference number
+        "normalize": "1"
+      }
       verify_transaction = Transaction.verify(payload, api=my_api)
 
       ```
-      *Arguments*
-
-        - Payload 
   
   - **verify transaction with xquery**
       *Usage*
       ```
+      payload = {
+        "flwref": "FLW-MOCK-6f52518a2ecca2b6b090f9593eb390ce", # unique reference for the transaction
+        "txref":"dummy", # merchants unique reference number
+        "last_attempt":"1", # retrieves the last transaction
+        "only_successful": "1" # retrieves only successful transaction
+        }
       xquery_verify = Transaction.verify_query(payload, api=my_api)
 
       ```
-      *Arguments*
 
   - **List all recurring transactions**
     *Usage*
     ```
-    from ravepaypysdk import Transaction
-
-    list_transactions  = Transaction.list_all_recurring(payload, api=my_api)
+    list_transactions  = Transaction.list_all_recurring(api=my_api)
     ```
-    *Arguments*
+
 
   - **List single recurring transactions**
     *Usage*
     ```
+    payload = {
+        "txId":"dummy" #add the required value for txId
+    }
     list_single_transaction = Transaction.list_single_recurring(payload, api=my_api)
     ```
-     *Payload*
 
-## Preauthorization 
-  Import the Preauthorization module
+## PreAuthorization
+  This module performs preauthorization transactions on the RavePay platform
+
+  *Usage*
+
+  Import the PreAuthorization module
   ```
-  from ravepaypysdk import Preauthorization
+  from ravepaypysdk import PreAuthorization
   ```
   - **Preauthorization Capture**
 
     *Usage*
     ```
-    preauthorize_capture = Preauthorization.capture(payload, api=my_api)
+    payload = {
+        "flwRef":"39448fhdhhfdhshshf" # add the required value
+    }
+    preauthorize_capture = PreAuthorization.capture(payload, api=my_api)
     ```
-    *Payload*
+
   
   - **Preauthorize Card transaction**
 
     *Usage*
     ```
-    preauthorize_card = Preauthorization.card(payload, api=my_api)
+    Payload is the same as direct charge card payment payload
+
+    preauthorize_card = PreAuthorization.card(payload, api=my_api)
 
     ```
-    *Payload*
+
 
   - **Preauthorize void or refund transactions**
 
     *Usage*
     ```
-    preauth_void_refund = Preauthorization.void_or_refund(payload, api=my_api)
+    payload = {
+        "flwRef":"dummy" # add the value from the capture response
+        "action": "refund or void" # select what action i.e refund or void
+    }
+    preauth_void_refund = PreAuthorization.void_or_refund(payload, api=my_api)
     ```
 
     *Payload*
 
  ## Validate Ravepay charges
-    Import the validation module
+   This module validates RavePay payment transactions.
 
-    ```
-    from ravepaypysdk import ValidateCharge
+   *Usage*
 
-    ```
+   Import the ValidateCharge module
 
-    This interface helps in validating transaction made through the RavePay platform
+   ```
+   from ravepaypysdk import ValidateCharge
+   payload = {
+        "transaction_reference": "222334304",
+        "otp": "12345"
+    }
+   ```
 
   - **Validate card transactions**
 
     *Usage*
     ```
+
     validate_card_transac = ValidateCharge.card(payload, api=my_api)
 
     ```
@@ -183,9 +261,11 @@ To start using the SDK ensure you have your public key and secret key instantiat
     ```
 
 ## Miscellanous 
-  Miscellanous includes getting the current forex rate and lists of banks supported by the RavePay Platform
+  This module gets the bank list and current forex rates
 
-  Import the bank module to get list of banks and current forex rates
+  *Usage*
+
+  Import the bank module
 
   ```
   from ravepaypysdk import Bank
@@ -202,8 +282,26 @@ To start using the SDK ensure you have your public key and secret key instantiat
 
     *Usage*
     ```
-    get_forex = Bank.get_forex(api=my_api)
+    payload = {
+     'origin_currency': 'USD',
+     'destination_currency': 'NGN',
+     'amount': '200'
+     }
+    get_forex = Bank.get_forex(payload, api=my_api)
     ```
+
+## Documentation
+API documentation for RavePay can be found [here](https://flutterwavedevelopers.readme.io/v1.0/reference#introduction)
+
+
+## Contributing
+
+Contributions are always welcomed to the project. Use Github Issue for requests.
+
+- Fork the project to your repository then clone it to your local machine.
+- Create a new branch and make the necessary enhancement to the features.
+- If the you wish to update an existing enhancement submit a pull request.
+- If you are unsure about certain areas in the project feel to ask for assistance.
