+++ ravepaypysdk/__init__.py
@@ -2,4 +2,4 @@ from .utils.encryption import encrypt_data, get_key
 from .utils.rave_utils import initialize_config, merge_dict, merge_url
 from .api import Api
 from .api_exceptions import ApiError
-from .services import PreAuthorization, Payment, ValidateCharge, Bank, Transaction
+from .resources import PreAuthorization, Payment, ValidateCharge, Bank, Transaction
